保存数据前判断

+++ cloud-service/src/main/java/com/jhmk/cloudservice/warnService/service/BasicInfoService.java
@@ -29,26 +29,26 @@ public class BasicInfoService {
     public void saveAndFlush(Rule rule) {
         String patient_id = rule.getPatient_id();
         String visit_id = rule.getVisit_id();
-//        List<BasicInfo> lessThanVisit_id = basicInfoRepService.findLessThanVisit_id(patient_id, visit_id);
-//        if (lessThanVisit_id.size() > 0) {
-//            basicInfoRepService.delete(lessThanVisit_id);
-//        }
-        basicInfoRepService.deleteByPatient_idAndVisit_id(patient_id, visit_id);
-        String pageSource = rule.getPageSource();
-        String warnSource = rule.getWarnSource();
-        String doctor_id = rule.getDoctor_id();
-        String doctor_name = rule.getDoctor_name();
-        String dept_code = rule.getDept_code();
-        //基本信息
-        BasicInfo basicInfo = new BasicInfo();
-        basicInfo.setPatient_id(patient_id);
-        basicInfo.setVisit_id(visit_id);
-        basicInfo.setDept_name(dept_code);
-        basicInfo.setDoctor_id(doctor_id);
-        basicInfo.setPageSource(pageSource);
-        basicInfo.setWarnSource(warnSource);
-        basicInfo.setDoctor_name(doctor_name);
-        basicInfoRepService.save(basicInfo);
+        BasicInfo byPatient_idAndVisit_id = basicInfoRepService.findByPatient_idAndVisit_id(patient_id, visit_id);
+        if (byPatient_idAndVisit_id == null || (byPatient_idAndVisit_id.getVisit_id() != null && byPatient_idAndVisit_id.getPatient_id() != null && byPatient_idAndVisit_id.getPageSource() != null)) {
+            basicInfoRepService.deleteByPatient_idAndVisit_id(patient_id, visit_id);
+            String pageSource = rule.getPageSource();
+            String warnSource = rule.getWarnSource();
+            String doctor_id = rule.getDoctor_id();
+            String doctor_name = rule.getDoctor_name();
+            String dept_code = rule.getDept_code();
+            //基本信息
+            BasicInfo basicInfo = new BasicInfo();
+            basicInfo.setPatient_id(patient_id);
+            basicInfo.setVisit_id(visit_id);
+            basicInfo.setDept_name(dept_code);
+            basicInfo.setDoctor_id(doctor_id);
+            basicInfo.setPageSource(pageSource);
+            basicInfo.setWarnSource(warnSource);
+            basicInfo.setDoctor_name(doctor_name);
+            basicInfoRepService.save(basicInfo);
+        }
+
     }
 
 }
